Parameterised OpenCL kernels.
Not all new macros are actually parameters intended to be changed. Rather, I tried to assign a meaningful name to most of magic numbers in the code.

The comments still need to be updated. Also, I removed a few kernel arguments that passed constant values, so the driver code need to be adapted to that.